requirements.txt
tqdm
numpy
torch==1.9.0
torchvision==0.10.0
easydict
pandas
jupyter
pyyaml
ipykernel
scipy
nltk
tensorboard
ftfy
regex
tqdm